Under construction and refactoring
