--- Wms.Web/Repositories/Abstract/IRepository.cs
@@ -1,6 +1,0 @@
-namespace Wms.Web.Repositories.Abstract;
-
-public interface IRepository
-{
-    
-}
