+++ Wms.Web/Repositories/Concrete/GenericRepository.cs
@@ -1,9 +1,9 @@
 using System.Linq.Expressions;
 using Microsoft.EntityFrameworkCore;
 using Wms.Web.Common.Exceptions;
-using Wms.Web.Repositories.Abstract;
+using Wms.Web.Repositories.Interfaces;
+using Wms.Web.Store.Common.Interfaces;
 using Wms.Web.Store.Entities.Interfaces;
-using Wms.Web.Store.Interfaces;
 
 namespace Wms.Web.Repositories.Concrete;
