+++ Wms.Web/Repositories/Interfaces/IRepository.cs
@@ -1,0 +1,6 @@
+namespace Wms.Web.Repositories.Interfaces;
+
+public interface IRepository
+{
+    
+}
