+++ Wms.Web/Services/Concrete/WarehouseService.cs
@@ -1,6 +1,6 @@
 using AutoMapper;
 using Wms.Web.Common.Exceptions;
-using Wms.Web.Repositories.Abstract;
+using Wms.Web.Repositories.Interfaces;
 using Wms.Web.Services.Abstract;
 using Wms.Web.Store.Entities;
 using Wms.Web.Services.Dto;
