+++ Wms.Web/Store.Common/WarehouseDbContext.cs
@@ -1,10 +1,10 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using Wms.Web.Store.Common.Interfaces;
 using Wms.Web.Store.Entities;
 using Wms.Web.Store.Entities.Interfaces;
-using Wms.Web.Store.Sqlite.Interfaces;
 
-namespace Wms.Web.Store.Sqlite;
+namespace Wms.Web.Store.Common;
 
 internal sealed class WarehouseDbContext : DbContext, IWarehouseDbContext
 {
