+++ Wms.Web/Store.Entities/Concrete/Palette.cs
