+++ Wms.Web/Store.Entities/Store.Entities.csproj
@@ -1,3 +1,1 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-</Project>
+<Project Sdk="Microsoft.NET.Sdk" />
