+++ Wms.Web/Store.Sqlite/Store.Sqlite.csproj
@@ -1,7 +1,5 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-   
-
+    
     <ItemGroup>
       <PackageReference Include="Microsoft.EntityFrameworkCore" />
     </ItemGroup>
