+++ Wms.Web/Wms.Web.sln
@@ -2,8 +2,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Services", "Services\Services.csproj", "{BA22245F-4BA6-4980-A611-4C54C0034B20}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Repositories", "Repositories\Repositories.csproj", "{288FA806-9E82-4CC1-BFB6-10EBDEB02D99}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Store.Common", "Store.Common\Store.Common.csproj", "{3B23E7A3-E62F-4A5B-875F-6CE4E8E42C02}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{FDE1330B-C59A-4DE2-9A45-803DF7B1B966}"
@@ -33,6 +31,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Store.Entities", "Store.Ent
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Store.Sqlite", "Store.Sqlite\Store.Sqlite.csproj", "{DDA88A2C-3FAD-437A-A63A-1CF1E45B2EE3}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Repositories", "Repositories\Repositories.csproj", "{2DEBC05A-512B-4957-A1E1-1C76B5B33960}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Store.Postgress", "Store.Postgress\Store.Postgress.csproj", "{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -43,10 +47,6 @@ Global
 		{BA22245F-4BA6-4980-A611-4C54C0034B20}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BA22245F-4BA6-4980-A611-4C54C0034B20}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BA22245F-4BA6-4980-A611-4C54C0034B20}.Release|Any CPU.Build.0 = Release|Any CPU
-		{288FA806-9E82-4CC1-BFB6-10EBDEB02D99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{288FA806-9E82-4CC1-BFB6-10EBDEB02D99}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{288FA806-9E82-4CC1-BFB6-10EBDEB02D99}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{288FA806-9E82-4CC1-BFB6-10EBDEB02D99}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3B23E7A3-E62F-4A5B-875F-6CE4E8E42C02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3B23E7A3-E62F-4A5B-875F-6CE4E8E42C02}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B23E7A3-E62F-4A5B-875F-6CE4E8E42C02}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -87,9 +87,28 @@ Global
 		{DDA88A2C-3FAD-437A-A63A-1CF1E45B2EE3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DDA88A2C-3FAD-437A-A63A-1CF1E45B2EE3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DDA88A2C-3FAD-437A-A63A-1CF1E45B2EE3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2DEBC05A-512B-4957-A1E1-1C76B5B33960}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2DEBC05A-512B-4957-A1E1-1C76B5B33960}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2DEBC05A-512B-4957-A1E1-1C76B5B33960}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2DEBC05A-512B-4957-A1E1-1C76B5B33960}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{147571E4-AD25-4EEA-A284-60210709FC80} = {0FD50018-3FE1-4EB1-B427-5AB10E4EF53B}
 		{60CAB536-5ECD-4047-9F2E-6909C768FAE3} = {0FD50018-3FE1-4EB1-B427-5AB10E4EF53B}
+		{B46BA97B-FEDD-45A5-A84F-BCF945C0982F} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{348BE50F-5CD4-4C80-9957-27931E66F2FD} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{5B4A03A4-A1DF-47DC-BDFF-50C5779821B9} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{97D27958-2932-48F5-A076-5C61EF25F278} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{6896BEDC-5733-4D43-998D-29D099249BCD} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{BA22245F-4BA6-4980-A611-4C54C0034B20} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{3B23E7A3-E62F-4A5B-875F-6CE4E8E42C02} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{56B9290A-01D2-421D-A7B3-998A39D16AE6} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{DDA88A2C-3FAD-437A-A63A-1CF1E45B2EE3} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{2DEBC05A-512B-4957-A1E1-1C76B5B33960} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
+		{9DD59541-6B1D-4D95-AE8A-5B941C42DEE1} = {18DAAB21-2FA0-4F7B-B52C-CCE5244F8540}
 	EndGlobalSection
 EndGlobal
